requirements.txt
gspread~=6.2.1
oauth2client~=4.1.3
openai~= 1.79.0
apscheduler~=3.11.0
python-telegram-bot~=22.1
serpapi~=0.1.5
chromadb~=1.0.9
PyYAML~=6.0.2
language-tool-python~=2.9.3
langchain[all]~=0.3.25
language_tool_python~=2.9.3
langchain-community~=0.3.24
google-search-results~=2.4.2
langchain-openai~=0.3.17
langchain_chroma~= 0.2.4

gspread~=6.2.1
oauth2client~=4.1.3
PyYAML~=6.0.2
openai~=1.79.0
langchain[all]~=0.3.25
langchain-openai~=0.3.17
chromadb~=1.0.9
apscheduler~=3.11.0
python-telegram-bot~=22.1
requests~=2.31.0
serpapi~=0.1.5
google-search-results~=2.4.2